Returns only the incident updates for the given incident.
Untill know all the updates of all the incident were returned when
calling /api/v1/incidents/{incident}/updates, but the goal of this
method is to return the updates for only one incident.

I've added a WHERE clause in order to filter the updates on the
incident.

Related to CachetHQ/Cachet/#2775