Show schedules till they are completed
Currently schedules are shown on the status page when they are in the status upcoming or in progress. But, when the scheduled time is in the past - it's hidden, even when it's still going on.

This PR fixes this issue by hiding the maintenance when it's in the status completed, or the completed time has passed.