Set the mail username optional in setup
During the setup process email informations were asked:
  - Driver
  - Host
  - Username
  - Password

In some situations the username is not useful because the Cachet's
host may be configured to forward email to a server.

The problem is the username was required, so we had to set a username
and then update the .env file to remove it.

To fix this problem, the mail username has been set to optional in
the setup. So if someone needs a username it still can use this field,
and otherwise people can let it empty.

See: #3244